Basic C# code generation of some basic features

+++ src/frontend/symbols.rs
@@ -9,14 +9,14 @@ impl Symbols {
         Self { table: Vec::new() }
     }
 
-    pub fn find_or_add(&mut self, span: &Span) -> u16 {
+    pub fn find_or_add(&mut self, span: &Span) -> u32 {
         for (idx, item) in self.table.iter().enumerate() {
             if span.slice_source() == item.slice_source() {
-                return idx as u16;
+                return idx as u32;
             }
         }
 
         self.table.push(span.clone());
-        self.table.len() as u16 - 1
+        self.table.len() as u32 - 1
     }
 }
